Create sitemesh for project p1: - add library sitemesh,jstl - create filter web xml, decorate config

+++ ToeicOnline/toeic-core-data/src/main/java/com/hochnt/core/data/daoimpl/AbstractDao.java
@@ -204,6 +204,8 @@ public class AbstractDao<ID extends Serializable, T> implements GenericDao<ID, T
         try {
             for (ID id : ids) {
                 T t = (T) session.get(persistenceClass, id);
+                if (t == null)
+                    continue;
                 session.delete(t);
                 count++;
             }
